backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
rapidfuzz==3.9.3
python-dotenv==1.0.1
anthropic==0.34.0
